+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## TBD
 
 - Added the (missing) 'prismatic oil' icon
+- Changed the Settings Menu Login url from '/login/email' to '/login' and added a custom polyfill to make the login page buttons work
 
 ## 0.8.40 (2026-03-13)